src/inference/requirements.txt
flask
tensorflow==2.17
pillow
requests
google-cloud-storage
google-cloud-bigquery
numpy
structlog
fuzzywuzzy
python-Levenshtein